refactor fargo3d laoder to use separate files
