+++ src/simdata/loaders/fargo3d/__init__.py
@@ -1,0 +1,1 @@
+from . loader import *
